+++ packages/typescript/configs/import.ts
@@ -1,0 +1,24 @@
+import type { Linter } from 'eslint';
+import importPlugin from 'eslint-plugin-import';
+
+const rules: Linter.RulesRecord = {
+  'import/extensions': [
+    'error',
+    'ignorePackages',
+    {
+      js: 'always',
+      jsx: 'always',
+      ts: 'always',
+      tsx: 'always',
+    },
+  ],
+};
+
+const config: Linter.Config = {
+  plugins: {
+    import: importPlugin,
+  },
+  rules,
+};
+
+export default config;